+++ commons-encrypt/src/main/java/com/hb0730/commons/encrypt/utils/KeyUtils.java
@@ -3,12 +3,23 @@ package com.hb0730.commons.encrypt.utils;
 import com.hb0730.commons.encrypt.exceptions.EncryptException;
 import com.hb0730.commons.encrypt.symmetry.SymmetricAlgorithm;
 import com.hb0730.commons.lang.CharUtils;
+import com.hb0730.commons.lang.StringUtils;
+import com.hb0730.commons.lang.nums.RandomUtils;
 
 import javax.crypto.KeyGenerator;
 import javax.crypto.SecretKey;
+import javax.crypto.SecretKeyFactory;
+import javax.crypto.spec.DESKeySpec;
+import javax.crypto.spec.DESedeKeySpec;
+import javax.crypto.spec.PBEKeySpec;
+import javax.crypto.spec.SecretKeySpec;
+import java.nio.charset.StandardCharsets;
+import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.security.Provider;
 import java.security.SecureRandom;
+import java.security.spec.InvalidKeySpecException;
+import java.security.spec.KeySpec;
 
 /**
  * 密钥工具类
@@ -17,6 +28,10 @@ import java.security.SecureRandom;
  * @since 2.1.0
  */
 public class KeyUtils {
+    private static final String PBE = "PBE";
+    private static final String DES = "DES";
+    private static final String DES_EDE = "DESede";
+
     /**
      * 生成 {@link SecretKey}，仅用于对称加密和摘要算法密钥生成
      *
@@ -70,6 +85,118 @@ public class KeyUtils {
             }
         }
         return keyGenerator.generateKey();
+    }
+
+    /**
+     * 生成 {@link SecretKey}，仅用于对称加密和摘要算法
+     *
+     * @param algorithm 算法
+     * @param keySpec   {@link KeySpec}
+     * @param provider  {@link Provider}
+     * @return {@link SecretKey}
+     */
+    public static SecretKey generateKey(String algorithm, KeySpec keySpec, Provider provider) {
+        SecretKeyFactory keyFactory = getSecretKeyFactory(algorithm, provider);
+        try {
+            return keyFactory.generateSecret(keySpec);
+        } catch (InvalidKeySpecException e) {
+            throw new EncryptException(e);
+        }
+    }
+
+    /**
+     * 生成 {@link SecretKey}，仅用于对称加密和摘要算法密钥生成
+     *
+     * @param algorithm 算法
+     * @param key       密钥，如果为{@code null} 自动生成随机密钥
+     * @param provider  {@link Provider}
+     * @return {@link SecretKey}
+     */
+    public static SecretKey generateKey(String algorithm, byte[] key, Provider provider) {
+        if (StringUtils.isBlank(algorithm)) {
+            throw new EncryptException("Algorithm is blank!");
+        }
+        SecretKey secretKey;
+        if (algorithm.startsWith(PBE)) {
+            secretKey = generatePBEKey(algorithm, (null == key) ? null : new String(key, StandardCharsets.UTF_8).toCharArray(), provider);
+        } else if (algorithm.startsWith(DES)) {
+            secretKey = generateDESKey(algorithm, key, provider);
+        } else {
+            // 其它算法密钥
+            secretKey = (null == key) ? generateKey(algorithm, provider) : new SecretKeySpec(key, algorithm);
+        }
+        return secretKey;
+    }
+
+    /**
+     * 生成 {@link SecretKey}
+     *
+     * @param algorithm DES算法，包括DES、DESede等
+     * @param key       密钥
+     * @param provider  {@link Provider}
+     * @return {@link SecretKey}
+     */
+    public static SecretKey generateDESKey(String algorithm, byte[] key, Provider provider) {
+        if (StringUtils.isBlank(algorithm) || !algorithm.startsWith(DES)) {
+            throw new EncryptException("Algorithm [{}] is not a DES algorithm!");
+        }
+        SecretKey secretKey;
+        if (null == key) {
+            secretKey = generateKey(algorithm, provider);
+        } else {
+            KeySpec keySpec;
+            try {
+                if (algorithm.startsWith(DES_EDE)) {
+                    // DESede兼容
+                    keySpec = new DESedeKeySpec(key);
+                } else {
+                    keySpec = new DESKeySpec(key);
+                }
+            } catch (InvalidKeyException e) {
+                throw new EncryptException(e);
+            }
+            secretKey = generateKey(algorithm, keySpec, provider);
+        }
+        return secretKey;
+    }
+
+    /**
+     * 生成PBE {@link SecretKey}
+     *
+     * @param algorithm PBE算法，包括：PBEWithMD5AndDES、PBEWithSHA1AndDESede、PBEWithSHA1AndRC2_40等
+     * @param key       密钥
+     * @return {@link SecretKey}
+     */
+    public static SecretKey generatePBEKey(String algorithm, char[] key, Provider provider) {
+        if (StringUtils.isBlank(algorithm) || !algorithm.startsWith(PBE)) {
+            throw new EncryptException("Algorithm [{}] is not a PBE algorithm!");
+        }
+        if (null == key) {
+            key = RandomUtils.randomString(32, true).toCharArray();
+        }
+        PBEKeySpec spec = new PBEKeySpec(key);
+        return generateKey(algorithm, spec, provider);
+    }
+
+    /**
+     * 获取{@link SecretKeyFactory}
+     *
+     * @param algorithm 对称加密算法
+     * @param provider  {@link Provider}
+     * @return {@link java.security.KeyFactory}
+     * @since 4.5.2
+     */
+    public static SecretKeyFactory getSecretKeyFactory(String algorithm, Provider provider) {
+
+        SecretKeyFactory keyFactory;
+        try {
+            keyFactory = (null == provider) //
+                    ? SecretKeyFactory.getInstance(getMainAlgorithm(algorithm)) //
+                    : SecretKeyFactory.getInstance(getMainAlgorithm(algorithm), provider);
+        } catch (NoSuchAlgorithmException e) {
+            throw new EncryptException(e);
+        }
+        return keyFactory;
     }
 
     /**